backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
sentence-transformers
qdrant-client
google-generativeai
PyPDF2
python-multipart
supabase
pandas
python-jose
pydantic
pydantic[email]